feat: initial withExecutionContext() implementation
Problem:

When a function calls another function, which then calls another function, ...
We will want to know the call stack in some cases to determine where the
final call originated from.

Solution:

Add a new withExecutionContext feature that allows us to keep track of the
call stack in a userfriendly way. The final result is a string that shows the call stack
(but only for the functions where the context was added).

See the tests for the best understanding of what is happening.

Details:
- In a function/method you can wrap the actual code to execute in an "execution context",
  providing some identifiers for the current context.
- Any nested call will have that previous execution context, and if they wrap their own code,
  their context will be appended to the previous.
- `getExecutionContext()` will provide the latest execution context so that something like
  telemetry can provide it in a metric.
- AsyncLocalStorage is used under the hood to prevent race conditions when it comes to a global context + async code

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>